add error check and split pg_lib from hook define file
